urukul
Module app

The RTIC application module

Modules

eth

Hardware task

ethernet_link

Software task

idle

Idle loop

init

Initialization function

settings_update

Software task

shared_resources 🔒

usb

Software task

Structs

Local 🔒

Local resources

Shared 🔒

User code end Shared resources

__rtic_internal_eth_Context

Execution context

__rtic_internal_ethernet_linkSharedResources

Shared resources ethernet_link has access to

__rtic_internal_ethernet_link_Context

Execution context

__rtic_internal_idleSharedResources

Shared resources idle has access to

__rtic_internal_idle_Context

Execution context

__rtic_internal_init_Context

Execution context

__rtic_internal_settings_updateLocalResources

Local resources settings_update has access to

__rtic_internal_settings_updateSharedResources

Shared resources settings_update has access to

__rtic_internal_settings_update_Context

Execution context

__rtic_internal_usbLocalResources

Local resources usb has access to

__rtic_internal_usbSharedResources

Shared resources usb has access to

__rtic_internal_usb_Context

Execution context

Statics

RTIC_ASYNC_MAX_LOGICAL_PRIO 🔒

Holds the maximum priority level for use by async HAL drivers.

__rtic_internal_ethernet_link_EXEC 🔒

__rtic_internal_settings_update_EXEC 🔒

__rtic_internal_usb_EXEC 🔒

Functions

ETH 🔒 ⚠

SDMMC 🔒 ⚠

Interrupt handler to dispatch async tasks at priority 1

eth 🔒

ethernet_link 🔒

idle 🔒

init 🔒

settings_update 🔒

usb 🔒