urukul

Struct Cfg

pub struct Cfg { /* private fields */ }

Implementations

impl Cfg

pub const ZERO: Self

Creates a new instance with a raw value of 0. Equivalent to [Self::new_with_raw_value(u24 :: new(0))].

pub const DEFAULT: Self

An instance that uses the default value 0x000700

pub const fn new() -> Self

๐Deprecated: Use Cfg::Default (or Cfg::DEFAULT in const context) instead

Creates a new instance of this struct using the default value

pub const fn raw_value(&self) -> u24

Returns the underlying raw value of this bitfield

pub const fn new_with_raw_value(value: u24) -> Cfg

Creates a new instance of this bitfield with the given raw value.

No checks are performed on the value, so it is possible to set bits that donโt have any accessors specified.

pub const fn builder() -> PartialCfg<0>

Creates a builder for this bitfield which ensures that all writable fields are initialized

pub const fn rf_sw(&self) -> u4

pub const fn with_rf_sw(&self, field_value: u4) -> Self

pub fn set_rf_sw(&mut self, field_value: u4)

pub const fn led(&self) -> u4

pub const fn with_led(&self, field_value: u4) -> Self

pub fn set_led(&mut self, field_value: u4)

pub const fn profile(&self) -> u3

pub const fn with_profile(&self, field_value: u3) -> Self

pub fn set_profile(&mut self, field_value: u3)

pub const fn io_update(&self) -> bool

pub const fn with_io_update(&self, field_value: bool) -> Self

pub fn set_io_update(&mut self, field_value: bool)

pub const fn mask_nu(&self) -> u4

pub const fn with_mask_nu(&self, field_value: u4) -> Self

pub fn set_mask_nu(&mut self, field_value: u4)

pub const fn clk_sel(&self) -> ClkSel

pub const fn with_clk_sel(&self, field_value: ClkSel) -> Self

pub fn set_clk_sel(&mut self, field_value: ClkSel)

pub const fn sync_sel(&self) -> bool

pub const fn with_sync_sel(&self, field_value: bool) -> Self

pub fn set_sync_sel(&mut self, field_value: bool)

pub const fn rst(&self) -> bool

pub const fn with_rst(&self, field_value: bool) -> Self

pub fn set_rst(&mut self, field_value: bool)

pub const fn io_rst(&self) -> bool

pub const fn with_io_rst(&self, field_value: bool) -> Self

pub fn set_io_rst(&mut self, field_value: bool)

pub const fn div_sel(&self) -> DivSel

pub const fn with_div_sel(&self, field_value: DivSel) -> Self

pub fn set_div_sel(&mut self, field_value: DivSel)